Moved and renamed members.py
